refactor: 리포트 리스트 스타일 수정
