+++ app/(anon)/reports/components/ReportDetailCard.tsx
