--- app/(anon)/reports/components/ex.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div></div>;
-}
